requirements.txt
fastapi
uvicorn
opencv-python
numpy
scipy
pillow
scikit-learn
jinja2
weasyprint
boto3
python-dotenv
pydantic
pytest
